requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
anthropic==0.39.0
httpx==0.27.2
python-dotenv==1.0.1
tenacity==9.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
langfuse>=2.44.0,<3.0.0
